third_party/blink/web_tests/svg/dynamic-updates/SVGFEBlendElement-svgdom-mode-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'mode' property of the SVGFEBlendElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS normalBlend.mode.baseVal is SVGFEBlendElement.SVG_FEBLEND_MODE_LIGHTEN
PASS multiplyBlend.mode.baseVal is SVGFEBlendElement.SVG_FEBLEND_MODE_NORMAL
PASS screenBlend.mode.baseVal is SVGFEBlendElement.SVG_FEBLEND_MODE_MULTIPLY
PASS darkenBlend.mode.baseVal is SVGFEBlendElement.SVG_FEBLEND_MODE_SCREEN
PASS lightenBlend.mode.baseVal is SVGFEBlendElement.SVG_FEBLEND_MODE_DARKEN
PASS normalBlend.mode.baseVal is SVGFEBlendElement.SVG_FEBLEND_MODE_NORMAL
PASS multiplyBlend.mode.baseVal is SVGFEBlendElement.SVG_FEBLEND_MODE_MULTIPLY
PASS screenBlend.mode.baseVal is SVGFEBlendElement.SVG_FEBLEND_MODE_SCREEN
PASS darkenBlend.mode.baseVal is SVGFEBlendElement.SVG_FEBLEND_MODE_DARKEN
PASS lightenBlend.mode.baseVal is SVGFEBlendElement.SVG_FEBLEND_MODE_LIGHTEN
PASS successfullyParsed is true

TEST COMPLETE
